Refactor Responses wire types into domain modules
